refactor(relax): extract socket driver helpers into socket_driver_utils
Move constants, structs, and helper functions from socket_driver.cpp's
anonymous namespace into a dedicated socket_driver_utils module. This
reduces socket_driver.cpp from 876 to 507 lines and improves its code
quality score from 47 to 61.

Update CMakeLists.txt and test/CMakeLists.txt to include the new source
file in both the library and the socket_driver_test target.
