+++ source/source_relax/test/CMakeLists.txt
@@ -22,6 +22,7 @@ AddTest(
   LIBS base device
   SOURCES socket_driver_test.cpp
           ../socket_driver.cpp
+          ../socket_driver_utils.cpp
           ../socket_frame.cpp
           ../socket_ipi.cpp
           ../../source_cell/update_cell.cpp
